Cambio de tipo de dato id, se agregan endpoints a TurnoController y se completa métodos a TurnoServece

+++ integradorClinica/src/main/java/com/backend/digitalhouse/integradorClinica/dto/salida/paciente/DomicilioSalidaDto.java
@@ -2,7 +2,7 @@ package com.backend.digitalhouse.integradorClinica.dto.salida.paciente;
 
 public class DomicilioSalidaDto {
 
-    private long id;
+    private Long id;
     private String calle;
     private int numero;
     private String localidad;
@@ -11,7 +11,7 @@ public class DomicilioSalidaDto {
     public DomicilioSalidaDto() {
     }
 
-    public DomicilioSalidaDto(long id, String calle, int numero, String localidad, String provincia) {
+    public DomicilioSalidaDto(Long id, String calle, int numero, String localidad, String provincia) {
         this.id = id;
         this.calle = calle;
         this.numero = numero;
@@ -19,11 +19,11 @@ public class DomicilioSalidaDto {
         this.provincia = provincia;
     }
 
-    public long getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
